step 3 is completed.

+++ BuzzNumber.java
@@ -1,10 +1,10 @@
 public class BuzzNumber {
 
-    static boolean isBuzz(int num){
+    static boolean isBuzz(long num){
         
 
         while (num >= 10) { // Reduce until single-digit
-            int lastDigit = num % 10;
+            long lastDigit = num % 10;
             if(lastDigit == 7) return true;
             else    num = num / 10 - 2 * lastDigit;
