+++ EvenOdd.java
@@ -1,10 +1,10 @@
 public class EvenOdd {
 
-     static boolean isNatural(int num){
+     static boolean isNatural(long num){
         return num >= 1? true : false;
      }
 
-     static boolean isEvenOdd(int num){ 
+     static boolean isEvenOdd(long num){ 
         return num % 2 == 0 ? true : false;
      }
 }
